kill the vfs_flags member in struct bhv_vfs
All flags are added to xfs_mount's m_flag instead.  Note that the 32bit
inode flag was duplicated in both of them, but only cleared in the
mount when it was not nessecary due to the filesystem beeing small enough.
Two flags are still required here - one to indicate the mount option
setting, and one to indicate if it applies or not.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29507a by kenmcd.

  move vfs_flags to xfs_mount.