ensure file size is logged on synchronous writes
Synchronous writes currently log inode changes before syncing
pages to disk.  Since the file size is updated on I/O completion
we wont be writing out the updated file size and if we crash the
file will have the wrong size.  This change moves the logging
after the syncing of the pages to ensure we log the correct file
size.
Merge of xfs-linux-melb:xfs-kern:29649a by kenmcd.

  ensure file size is logged on synchronous writes