kill BMAPI_DEVICE
There is no reason to go into the iomap machinery just to get the
right block device for an inode.  Instead look at the realtime flag
in the inode and grab the right device from the mount structure.

I created a new helper, xfs_find_bdev_for_inode instead of opencoding
it because I plan to use it in other places in the future.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29680a by kenmcd.

  kill BMAPI_DEVICE