kill BMAPI_UNWRITTEN
There is no reason to go through xfs_iomap for the BMAPI_UNWRITTEN
because it has nothing in common with the other cases.  Instead
check for the shutdown filesystem in xfs_end_bio_unwritten and perform
a direct call to xfs_iomap_write_unwritten (which should be renamed
to something more sensible one day)


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29681a by kenmcd.

  kill BMAPI_UNWRITTEN