fix valid but harmless sparse warning
The new xlog_recover_do_reg_buffer checks call be16_to_cpu on di_gen
which is a 32bit value so sparse rightly complains.  Fortunately the
warning is harmless because we don't care for the value, but only
whether it's non-NULL.  Due to that fact we can simply kill the endian
swaps on this and the previous di_mode check entirely.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29709a by kenmcd.

  fix valid but harmless sparse warning