simplify vn_revalidate
No need to allocate a bhv_vattr_t on stack and call xfs_getattr to
update a few fields in the Linux inode from the XFS inode, just
do it directly.

And yes, this function is in dire need of a better name and prototype,
I'll do in a separate patch, though.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29713a by kenmcd.

  simplify vn_revalidate