kill unnessecary ioops indirection
Currently there is an indirection called ioops in the XFS data I/O
path.  Various functions are called by functions pointers, but there
is no coherence in what this is for, and of course for XFS itself
it's entirely unused.  This patch removes it instead and significantly
reduces source and binary size of XFS while making maintaince easier.
Merge of xfs-linux-melb:xfs-kern:29737a by kenmcd.

  kill unnessecary ioops indirection