xlog_rec_header/xlog_rec_ext_header endianess annotations
Mostly trivial conversion with one exceptions:  h_num_logops was kept
in native endian previously and only converted to big endian in
xlog_sync, but we always keep it big endian now.  With todays cpus
fast byteswap instructions that's not an issue but the new variant
keeps the code clean and maintainable.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29821a by kenmcd.

  xlog_rec_header/xlog_rec_ext_header endianess annotations