Put back the QUEUE_ORDERED_NONE test in the barrier check.
Put back the QUEUE_ORDERED_NONE test which caused us grief in sles when it was taken out
as, IIRC, it allowed md/lvm to be thought of as supporting barriers when they weren't
in some configurations.
This patch will be reverting what went in as part of a change for
the SGI-pv 964544 (SGI-Modid: xfs-linux-melb:xfs-kern:28568a).
Merge of xfs-linux-melb:xfs-kern:29882a by kenmcd.

  Put back the QUEUE_ORDERED_NONE test in the barrier check.