Fix sparse warning in xlog_recover_do_efd_trans.
Sparse trips over the locking order in xlog_recover_do_efd_trans()
when xfs_trans_delete_ail() drops the ail lock. Because the unlock
is conditional, we need to either annotate with a "fake unlock" or
change the structure of the code so sparse thinks the function
always unlocks.

Reordering the code makes it simpler, so do that.
Merge of xfs-linux-melb:xfs-kern:30003a by kenmcd.

  Fix sparse warning in xlog_recover_do_efd_trans.