Fix transaction overrun during writeback.
Prevent transaction overrun in xfs_iomap_write_allocate() if we
race with a truncate that overlaps the delalloc range we were
planning to allocate.

If we race, we may allocate into a hole and that requires block
allocation. At this point in time we don't have a reservation for
block allocation (apart from metadata blocks) and so allocating
into a hole rather than a delalloc region results in overflowing
the transaction block reservation.

Fix it by only allowing a single extent to be allocated at a
time.
Merge of xfs-linux-melb:xfs-kern:30005a by kenmcd.

  Only allow xfs_iomap_write_allocate to allocate a single extent at a
  time to prevent races with truncate from causing unreserved allocation
  and hence transaction overruns.