Remove bogus assert,
lachlan says:
> This assert is bogus.  We can have a forced shutdown occur between the check for
the XLOG_FORCED_SHUTDOWN and the ASSERT.  Also, the logging system shouldn't care
about the state of XFS_FORCED_SHUTDOWN, it should only check XLOG_FORCED_SHUTDOWN.
The logging system has it's own forced shutdown flag so, for the case of a forced
shutdown that's not due to a logging error, we can flush the log.  I suggest
> removing the ASSERT.
Merge of xfs-linux-melb:xfs-kern:30029a by kenmcd.

  remove bogus assert