Fix broken inode cluster setup.
The radix tree based inode caches did away with the inode cluster hashes,
replacing them with a bunch of masking and gang lookups on the radix tree.

This masking got broken when moving the code to per-ag radix trees and
indexing by agino # rather than straight inode number. The result is
clustered inode writeback does not cluster and things can go extremely
slowly when there are lots of inodes to write.

Fix it up by comparing the agino # of the inode we just looked up
to the index of the cluster we are looking for.

Signed-off-by: Dave Chinner <dgc@sgi.com>
Tested-by:  Torsten Kaiser <just.for.lkml@googlemail.com>
Merge of xfs-linux-melb:xfs-kern:30033a by kenmcd.

  Make the cluster lookup use the agino rather than full inode number
  so that clusters are correctly built.