Fix xfs_ichgtime()s broken usage of I_SYNC
The recent I_LOCK->I_SYNC changes mistakenly changed xfs_ichgtime to
look at I_SYNC instead of I_LOCK. This was incorrect and prevents newly
created inodes from moving to the dirty list. Change this to the correct
check which is for I_NEW, not I_LOCK or I_SYNC so that behaviour is
correct.
Merge of xfs-linux-melb:xfs-kern:30204a by kenmcd.

  Use I_NEW in xfs_ichgtime instead of I_SYNC so that new inodes are
  moved to the dirty list correctly.