Put the correct offset in dirent d_off
The recent filldir regression fix was not putting the correct
d_off in each dirent. This was resulting in incorrect cookies being
passed to dmapi ioctls and the wrong offset appearing in the dirents.
readdir was unaffected as the filp->f_pos was being updated with the
correct offset and this was being written into the last dirent in
each buffer. Fix the XFS code to do the right thing.
Merge of xfs-linux-melb:xfs-kern:30240a by kenmcd.

  Update filp->f_pos correctly now that we pass the current offset with
  each dirent to filldir.