Initialise current offset in xfs_file_readdir correctly
After reading the directory contents into the temporary buffer, we
grab each dirent and pass it to filldir witht eh current offset of
the dirent. The current offset was not being set for the first
dirent in the temporary buffer, which coul dresult in bad offsets
being set in the f_pos field result in looping and duplicate entries
being returned from readdir.
Merge of xfs-linux-melb:xfs-kern:30282a by kenmcd.

  Initialise the curr_offset correctly for the first dirent in each temp
  buffer during readdir.