prevent panic during log recovery due to bogus operation header length
A problem was reported where a system panicked in log recovery due
to a corrupt log record.  The cause of the corruption is not known
but this change will at least prevent a crash for this specific
scenario.  Log recovery definitely needs some more work in this area.
Merge of xfs-linux-melb:xfs-kern:30318a by kenmcd.

  prevent panic during log recovery due to bogus operation header length