[PATCH 4/4] use generic_permission
Now that all direct caller of xfs_iaccess are gone we can kill
xfs_iaccess and xfs_access and just use generic_permission with a
check_acl callback.  This is required for the per-mount read-only
patchset in -mm to work properly with XFS.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:30370a by kenmcd.

  use generic_permission