Make xfs_ail_check check less by default
Checking the entire AIL on every insert and remove is
prohibitively expensive - the sustained sequntial create rate
on a single disk drops from about 1800/s to 60/s because of
this checking resulting in the xfslogd becoming cpu bound.

By default on debug builds, only check the next and previous
entries in the list to ensure they are ordered correctly.
If you really want, define XFS_TRANS_DEBUG to use the old
behaviour.
Merge of xfs-linux-melb:xfs-kern:30372a by kenmcd.

  Make exhaustive AIL checking only occur when XFS_TRANS_DEBUG
  is defined to prevent excessive performance loss on normal
  debug kernels.