keep i_nlink updated and use proper accessors
To get the read-only bind mounts in -mm to work correctly with XFS we
need to call the drop_nlink and inc_nlink helpers to monitor the link
count.  Add calls to these to xfs_bumplink and xfs_droplink and stop
copying over di_nlink to i_nlink in xfs_validate_fields and vn_revalidate.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:30392a by kenmcd.

  keep i_nlink updated and use proper accessors