Factor xfs_itobp() and xfs_inotobp().
The only difference between the functions is one passes an
inode for the lookup, the other passes an inode number.
However, they don't do the same validity checking or set
all the same state on the buffer that is returned yet
they should.

Factor the functions into a common implementation.
Merge of xfs-linux-melb:xfs-kern:30500a by kenmcd.

  Factor xfs_itobp() and xfs_inotobp() to use a common
  implementation.