kill xfs_get_dir_entry
Instead of of xfs_get_dir_entry use a macro to get the xfs_inode from
the dentry in the callers and grab the reference manually.

Only grab the reference once as it's fine to keep it over the dmapi
calls.  (And even that reference is actually superflous in Linux but
I'll leave that for another patch)

Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:30531a by kenmcd.

  kill xfs_get_dir_entry