Update c/mtime correctly on truncates
XFS changes the c/mtime of an inode when truncating it to the same
size. The c/mtime is only supposed to change if the size is changed.
Not to be confused with ftruncate, where the c/mtime is supposed to
be changed even if the size is not changed.

The Linux VFS encodes this semantic difference in the flags it sends
down to ->setattr, which XFS currently ignores. We need to make XFS
pay attention to the VFS flags and hence Do The Right Thing.
Merge of xfs-linux-melb:xfs-kern:30536a by kenmcd.

  Only change the c/mtime on truncate when the file size does
  not change if asked to by the VFS.