Use power-of-2 sized buffers to reduce overhead
Now that the ktrace_enter() code is using atomics,
the non-power-of-2 buffer sizes - which require modulus
operations to get the index - are showing up as using
substantial CPU in the profiles.

Force the buffer sizes to be rounded up to the nearest
power of two and use masking rather than modulus operations
to convert the index counter to the buffer index. This
reduces ktrace_enter overhead to 8% of a CPU time, and
again almost halves the trace intensive test runtime.
Merge of xfs-linux-melb:xfs-kern:30538a by kenmcd.

  Use power-of-2 buffer sizes so we can use more efficient
  masking operations to get the buffer index.