Fix superblock features2 field alignment problem
Due to the xfs_dsb_t structure not being 64 bit aligned,
the last field of the on-disk superblock can vary in location
This causes problems when the filesystem gets moved to a
different platform, or there is a 32 bit userspace and 64 bit
kernel.

This patch detects the defect at mount time, logs a warning
such as:

XFS: correcting sb_features alignment problem

in dmesg and corrects the problem so that everything is OK.
it also blacklists the bad field in the superblock so it does
not get used for something else later on.
Merge of xfs-linux-melb:xfs-kern:30539a by kenmcd.

  Detect and correct the features2 field of the superblock
  being misaligned. Blacklist the misaligned field so it
  does not get reused in future.