If you mount an XFS filesystem with no mount options at all, then the "ikeep" option is set rather than "noikeep".
This regression was introduced in 970451.

With no mount options specified, xfs_parseargs() does the following:

	int			ikeep = 0;

	args->flags |= XFSMNT_BARRIER;
	args->flags2 |= XFSMNT2_COMPAT_IOSIZE;

	if (!options)
		goto done;

It only sets the above two options by default and before, it also used to
set XFSMNT_IDELETE by default.

If options are specified, then

	if (!(args->flags & XFSMNT_DMAPI) && !ikeep)
		args->flags |= XFSMNT_IDELETE;

is executed later on which is skipped by the "goto done;" above.

The solution is to invert the logic.
Merge of xfs-linux-melb:xfs-kern:30590a by kenmcd.

  Change the *_IDELETE flags to *_IKEEP, and flip the logic as necessary.