Remove superflous xfs_readsb call in xfs_mountfs.
When xfs_mountfs is called by xfs_mount xfs_readsb was called 35 lines
above unconditionally, so there is no need to try to read the superblock
if it's not present.  If any other port doesn't have the superblock
read at this point it should just call it directly from it's xfs_mount
equivalent.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:30603a by kenmcd.

  Remove superflous xfs_readsb call in xfs_mountfs.