Prevent xfs_bmap_check_leaf_extents() from referencing unmapped memory.
While investigating the extent corruption bug I ran into this bug in
debug only code.  xfs_bmap_check_leaf_extents() loops through the leaf
blocks of the extent btree checking that every extent is entirely
before the next extent.  It also compares the last extent in the previous
block to the first extent in the current block when the previous block
has been released and potentially unmapped.  So take a copy of the last
extent instead of a pointer.  Also move the last extent check out of the
loop because we only need to do it once.
Merge of xfs-linux-melb:xfs-kern:30718a by kenmcd.

  Prevent xfs_bmap_check_leaf_extents() from referencing unmapped memory.