Remove the xlog_ticket allocator
The ticket allocator is just a simple slab implementation
internal to the log. It requires the icloglock to be held
when manipulating it and this contributes to contention
on that lock.

Just kill the entire allocator and use a memory zone instead.
While there, allow us to gracefully fail allocation with ENOMEM.
Merge of xfs-linux-melb:xfs-kern:30771a by kenmcd.

  Remove the xlog_ticket allocator and replace it with a memory zone.