Ensure "both" features2 slots are consistent
Since older kernels may look in the sb_bad_features2 slot for
flags, rather than zeroing it out on fixup, we should make it
equal to the sb_features2 value.

Also, if the ATTR2 flag was not found prior to features2
fixup, it was not set in the mount flags, so re-check after the
fixup so that the current session will use the feature.

Also fix up the comments to reflect these changes.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>
Merge of xfs-linux-melb:xfs-kern:30778a by kenmcd.

  Ensure "both" features2 slots are consistent and the ATTR2
  mount flag is set properly.