Clean up quotamount error handling.
xfs_qm_mount_quotas() returns an error status that is ignored.
If we fail to mount quotas, we continue with quota's turned
off, which is all handled inside xfs_qm_mount_quotas(). Mark
it as void to indicate that errors need not be returned to the
callers.
Merge of xfs-linux-melb:xfs-kern:30788a by kenmcd.

  Clean up quotamount error handling.