Catch errors when turning off quotas.
When turning off quota, we need to write various transactions
to the log to ensure that they are cleanly removed in the case
of a crash. We need to check that the transactions hit the disk
correctly. If we fail to write the final quota off transaction,
we are corrupt in memory and so the only option is to shut the
filesystem down at this point.
Merge of xfs-linux-melb:xfs-kern:30790a by kenmcd.

  Catch errors when turning off quotas.