Propagate errors from xfs_trans_commit().
xfs_trans_commit() can return errors when there are problems
in the transaction subsystem. They are indicative that the entire
transaction may be incomplete, and hence the error should be
propagated as there is a good possibility that there is something
fatally wrong in the filesystem. Catch and propagate or warn
about commit errors in the places where they are currently ignored.
Merge of xfs-linux-melb:xfs-kern:30795a by kenmcd.

  Propagate errors from xfs_trans_commit()