Clean up xfs_alloc_search_busy() return values.
xfs_alloc_search_busy() returns an index into the busy
array if the extent was found in the array. This is never checked,
and the xfs_alloc_search_busy() does a log force to prevent
reuse of the extent before the free transaction hits the disk.
Hence the return value is useless. Declare the function void
and remove the slot number from the tracing as well.
Merge of xfs-linux-melb:xfs-kern:30796a by kenmcd.

  Clean up xfs_alloc_search_busy() return values.