Make xfs_alloc_compute_aligned() void.
xfs_alloc_compute_aligned() returns a value based on a comparison
of the computed extent length and the minimum length allowed. This
is only used by some callers - the other four return parameters
are used more often. Hence move the comparison to the code that
actually needs to do it and make xfs_alloc_compute_aligned() a
void function.
Merge of xfs-linux-melb:xfs-kern:30797a by kenmcd.

  Make xfs_alloc_compute_aligned() void.