Ensure errors from xfs_bdstrat() are correctly checked.
xfsbdstrat() is declared to return an error. That is never checked
because the error is propagated by the xfs_buf_t that is passed
through the function.

Mark xfsbdstrat() as returning void and comment the prototype
on the methods needed for error checking.
Merge of xfs-linux-melb:xfs-kern:30823a by kenmcd.

  Ensure errors from xfs_bdstrat() are correctly checked.